requirements.txt
alpaca-py
requests
schedule
pytz
pandas
